requirements.txt
flask
google-generativeai=0.8.3